refactor: colocate Android alert handling
