+++ src/backend.ts
@@ -1,5 +1,5 @@
 import type { AndroidSnapshotBackendMetadata } from './platforms/android/snapshot-types.ts';
-import type { AlertInfo } from './alert-contract.ts';
+import type { AlertAction, AlertInfo } from './alert-contract.ts';
 import type { AppsFilter } from './commands/app-inventory-contract.ts';
 import type {
   Point,
@@ -111,7 +111,7 @@ export type BackendClipboardTextResult = {
   text: string;
 };
 
-export type BackendAlertAction = 'get' | 'accept' | 'dismiss' | 'wait';
+export type BackendAlertAction = AlertAction;
 
 export type BackendAlertInfo = AlertInfo;
